chore(jsii): improve locality of JSII9998 error
When a declaration cannot be represented in a jsii assembly, the error
message was localized on the whole declaration, but it is a better
experience if the message is only labelled on the declaration's name,
when that can be figured out - after all, the name is what will not be
visible in other languages, as the rest of the declaration is really
just an implementation detail that otherwise does not get conveyed into
foreign languages either.

This would result in more compact warnings generated for this particular
offense, which would be a quality-of-life improvement (can see more
diagnostic messages on a single screen, given a fixed size).