doc(decision): go callback formalism
Documents the decision for the userland formalism for go overrides. It
explains the way dynamic dispatch can be achieved in go, and how
overrides are to be declared.

The runtime mechanism itself derives directly from the userland syntax,
and is relaitvely straight-forward (as it does not differ significantly
from the mechanisms found in other language runtime libraries).