feat(go): run `go fmt` and `go build` after code generation
To ensure the generated go code is formatted and compiles by executing `go fmt`
and `go build` on the output directory.

In order to be able to do that within mono-repos (and in the jsii repo itself), we need to be able to resolve local dependencies (same as all other languages).

Resolves #2463