feat(runtime): use a dedicated file descriptor for sync IPC (#2411)
This replaces the `@jsii/runtime` package entry point so that it creates
a child `node` process with two new pipe file descriptors (at FD3 and
FD4). This means the `@jsii/runtime` now has two processes instead of
one:

1. A (new) "wrapper" process
2. The "core" process

> The "wrapper" process would not be necessary if our host runtimes
> consistently supported spawning child processes with additional open
> file descriptors, which is not possible in at least **Java** and **C#**...
> It would however be possible to bypass the "wrapper" process entirely
> in languages with the right APIs (at which point, the host app would
> interact with the child process like the "wrapper" would, and overhead
> would be reduced). This optimisation has however been intentionally
> excluded from this PR for now.

---

The `@jsii/runtime` core process now uses `FD#3` for synchronous reading
of requests sent by the host to the `@jsii/kernel`, synchronously write
responses back to the host process. This alleviates the need to resort to
hot-polling around `EAGAIN` errors that are symptomatic of attempting
blocking operations on a non-blocking file descriptor (aws/aws-cdk#12069).
This works as the non-standard (`>2`) file descriptors are not managed
by `libuv`, and hence are not made non-blocking at the start of `node`,
allowing blocking operations to be reliably used with those descriptors.

The `@jsii/runtime` core process was updated to use those two new file
descriptors instead of directly using `STDIN` and `STDOUT`, allowing
the **JavaScript** code to naturally use `process.stdout` and
`process.stderr` (as well as all other APIs that eventually use the same
file descriptors, such as `console.log` and `console.error`) without
causing corruption of the host-kernel communication channel as was
previously the case (#2406).

Data written directly to the `@jsii/runtime` core process' `FD#1`
(`STDOUT`) and `FD#2` (`STDERR`) is intercepted by the wrapper process,
and encoded in JSON objects before being sent to the host application
via the wrapper process' `STDERR` pipe. The host app is now reponsible
to maintain a thread that reads from the wrapper's `STDERR` pipe, decode
the JSON objects, then write the data to the appropriate host process'
STDIO stream to emulate direct operations as closely as possible.

---

Finally, an explicit `exit` message was added to the `@jsii/runtime`
protocol, as it is difficult to ensure the end-of-file is properly reached
on the `STDIN` IPC stream across all supported platforms. This has been
particularly challenging with `C#` on Windows, where the IPC pipes in
the parent `node` process are exposed as a duplex `Socket`s that do not
allow half-open mode. This makes it impossible to close only the writing
side of the socket without fully closing it (causing `EPIPE`), and as a
consequence those `Socket`s remain on the event loops, preventing the
`node` VM from "naturally" terminating... The `exit` command gives a
signal when `process.exit` can be called to manually terminate the VM.

---

Fixes aws/aws-cdk#12069
Fixes #2406

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0