fix(go): invalid major version suffix for submodules and jsii init
The major version suffix introduced in #2507 was added at the end of the full module name, but for submodules, the suffix is needed only after the root package name.

Additionally, the `_init_` import failed to include the major version suffix as well.

In order to cover these cases in tests, change the version of `jsii-calc` to `3.20.1` so that it will be a MV larger then 2.0.

Fixes #2507