fix(java): JsiiRuntime.ErrorStreamSink does not respond to being interrupted
When the thread is interrupted, it should understand the VM is asking it to shut down,
and stop polling the child process' STDERR for more data. Previously, the thread did
not check Thread.interrupted(), and hence was ignoring requests to terminate, resulting
in ugly errors during clean-up.

This adds the necessary check for interrupt status, and stops the error sink as
asked by the VM.

Fixes #2533