fix(rosetta,pacmak): TypeError in node 10 with --experimental-worker (#2554)
isaacs/node-graceful-fs#204 prevents `graceful-fs` (a dependency of
`fs-extra`) from loading in a worker thread when using node < 12. The
issue affects only version `4.2.5`, so pinning both `jsii-rosetta` and
`jsii-pacmak` (both of which may use worker threads when available) to
the last known good version: `4.2.4`, until a fixed release of
`graceful-fs` is released.