chore(go): runtime-wide type registration (#2567)
The previous convention of passing of `implementationMap` values at the
call sites for value-returning calls (`Get`, `Invoke`, and their static
counterparts) did not provide sufficient information to properly convert
certain values (JavaScript returning a descendant of the declared return
type, decoding enum members, ability to represent values that implement
multiple interfaces in JavaScript, ...).

In order to address this shortcomings, this mechanism was replaced with
a new `func init(){}` block emitted in each package, which is
responsible for registering the generated types with the runtime
library. This allows the `client` to determine the correct go type to
use when converting arbitrary values returned from JavaScript, as it has
visibility on the complete set of available types, by fully qualified
name.

This minimally changes the conversion of JavaScript values into go
values, so as to minimize the footprit of this change. Of course,
subsequent effort will be made to fully leverage the potential of the
extended information (i.e: correctly processing enum members, ...).

Fixes #2561

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0