refactor(go): re-organize the go runtime
Split the go runtime into multiple submodules, in an attempt to ensure
we do not horizontally couple components that should not be coupled to
each other, have a cleaner API for each component (since unexported
members are simply not visible). This also improves the quality of code
completion is improved as a result, too.