chore: copy golang runtime to dist for publishing
The publisher expects golang code to reside under the `go` directory in the build artifact. To that end, add a "package" script to the `@jsii/go-runtime` module which copies the go runtime code to `dist/go`. This is then collected during mono-repo packaging.

Remove the go publishing step from the github workflow as it is now handled by the delivlib pipeline.

Fixes #2579