feat(go): require go 1.16, use native embed
Instead of generating code to build byte slices with the embedded
assets, upgraded the baseline go requirement to go 1.16, and use the new
native embed feature released as part of this go release.

This includes the update to `jsii/superhain` to the correct go release,
updates to the `.github/workflows/main.yml` workflow to ensure the
correct go version is installed in PR validation builds, and the changes
in `jsii-pacmak` and `@jsii/go-runtime` to leverage the feature.

See: https://pkg.go.dev/embed