fix(python): accept Sequence[T] for array parameters
Instead of typing array (list) parameters as `typing.List[T]`, using the
more flexible `typing.Sequence[T]` type, as recommended by the `typing`
package documentation for this use-case.

Using `List[T]` is inconvenient as it is invariant ([ref]), and requires
very intentional typing, which elads to boilerplate. `Sequence[T]` is
variant, which removes the need for boilerplate, as long as all items
in the sequence match the expected type.

[ref]: https://mypy.readthedocs.io/en/latest/common_issues.html#variance

This was reported in aws/aws-cdk#13203