refactor(go): hide API keys from the kernel API
Those need not be set on each site that makes a kernel call. This leads
to possible errors. Instead, the API call name is no centralized by the
`kernel` package, which sets it "behind the covers".

In order for the naming to make sense with respects to go naming
conventions, I renamed:
- `type kernelRequester struct` to `kernelRequest`
- `type kernelRequest interface` to `kernelRequester`
- `type kernelResponder struct` to `kernelResponse`
- `type kernelResponse interface` to `kernelResponder`

The `kernelRequest` struct now contains the `API` field definition, and
all the client APIs that end up calling `client.request` compose a
`kernelRequest` and their "request properties type" into a single
struct, for easy and straight-forward marshaling.