fix(go): nested types are not namespaced
Nested types were generated in the same package as they parent, without
any namespacing additions, meaning that if two types in the same package
have nested types with the same name, the generated code would be
invalid.

This namespaces the nested types in go by prefixing their names with
their nesting type's name, using an `_` delimiter, which is the same as
what is done for static methods.

Also added a validation in the `jsii` compiler that prohibits that a
nested type and a static method share the same PascalCase
transformation, as this would result in conflicts in Go, but also in C#.

Fixes #2649