feat(go): packageName and versionSuffix (#2687)
Add two features to the `go` targets config in package.json:

- `packageName` can be used to customize the name of the go package (previously this was always derived from the javascript name).
- `versionSuffix` can be used to append a suffix to the version number (e.g. `-devpreview`).

Resolves #2632

This is needed in order to control the name of the AWS CDK go bindings so they are the same between v1 and v2. Currently the v1 name would have been `monocdk` and v2 `awscdklib` and we want both to actually be `awscdk` per the decision in [this RFC](https://github.com/aws/aws-cdk-rfcs/blob/master/text/0079-cdk-2.0.md#releasing-to-package-managers).