fix(go): embedded bases may conflict with member names
When we generate proxy structs in Go (for classes or interfaces), we embed the base types (base class or interfaces) in the struct. If the base type has a name that is the same as a member of the hosting type, there will be a conflict.

To solve this, we create a unique type alias for each base interface we want to embed.

Fixes #2702