feat(superchain): install both Node 10 and Node 14
We want to switch to Node 14 for CDK v2, and so we need Node 14 be
present in the Superchain image.

At the same time, we want to continue testing v1 against Node 10, so
we ALSO need Node 10 present in the image.

Use NVM to install both versions in parallel and providing a command
to switch between the two.

Node 10 is the default, build scripts can switch to Node14 by doing
either:

```
$ nvm use 14
$ nvm exec 14 <command>
```