fix(go): protected property accessors missing (#2738)
Not sure if the original issue was referring to protected *methods* as well as properties, but turns out protected methods are already treated as public one. 

So this PR just fixes the handling of protected properties, as well as adjusts the compliance tests.  

Fixes https://github.com/aws/jsii/issues/2673

---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0