feat(superchain): also build a superchain image with Node 14
For CDKv2, we want to require (and test) against a newer version of
Node. We've attempted to build multiple Node versions into a single
Docker image, but because we cannot control the entrypoint there is no
reliable way of switching between the Node versions at execution time.

Therefore, we will now build two versions of the "superchain" image.
The Node version built into each is controlled by the `NODE_VERSION`
build argument.

Workflow has been changed to build and push the following images:

* Nightlies:
    * `jsii/superchain:node10-nightly`
    * `jsii/superchain:node14-nightly`
    * `jsii/superchain:nightly` will remain an alias for `:node10-nightly`
* Releases:
    * `jsii/superchain:node10`
    * `jsii/superchain:node14`
    * `jsii/superchain:latest` will remain an alias for `:node10`.

(Multiple images are built in a `for` loop instead of having parallel
workflows because as far as I know there's no way to parameterize the
rather complex GitHub Action Workflow, so the alternative would be
copy/pasting which I'd rather not do).