fix(go): unable to set an array of interfaces/enums
Contents of map objects were not correctly de-serialized upon receiving
them from the node process: the re-hydratation was not recursive, and
also hit a bug in array re-hydratation when the receiving pointer does
not point to a slice type (i.e: it is an `interface{}`).

The runtime is now correctly re-hydrating recursively through maps, and
arrays can be correctly re-hydrated into `interface{}` values.

Fixes #2686