fix(go): panic on callback: using *<interface> as <interface>
The callback handling logic was incorrectly handling "pure" interface
parameters. The reflection piece would create a new argument value in
preparation for the callback, however for interfaces, it would leave the
pointer to the interface type (as returned by `reflect.New`) instead of
indirecting to the interface itself as appropriate.

Fixes #2793