fix: crash when assembly originator key file is missing
Instead of failing when the assembly originator key file is not found,
warn about it being missing and how the assembly will not be strong-name
signed as a result.

This is to be more compatible with the previously "broken" behavior,
where this file could be omitted or not actually used.