fix(rosetta): fails on "Debug Failure"
In cases where a literate source file was missing, the substitution
value was not valid TypeScript, which could cause the comipler to fail
on an opaque error (`Debug Failure`).

This changes the substitution text to make it more clear that an example
could not be rendered because the source for it could not be found, and
also makes sure to rescue the `Debug Failure` where it is emitted from,
that is in the call to `program.getDeclarationDiagnostics`.

Both of this changes result in greater chances of transliteration
success.

Related to cdklabs/jsii-docgen#369