feat(superchain): arm64 support
Introducing ARM64 support for the `jsii/superchain` docker image using
the multi-arch capabilities of `docker buildx`. This prompted a couple
of changes in the image, which should be fine for the majority of
use-cases, but is a breaking change in certain edge cases...

The new image is based off `debian:10-slim`, as unfortunately, binary
distributions of some of the image contents are not available for ARM64
platform on the RHEL/CentOS-based distributions (which `amazonlinux:2`
is). This means packages now get installed from `apt` instead of `yum`.

Image testing (by building `jsii` within the image) is now part of the
image build process itself, instead of being separated, as this allows
testing on all supported architectures (which is otherwise somewhat more
difficult to achieve).

BREAKING CHANGE: the `jsii/superchain` image is no longer based off
`amazonlinux:2`, and instead is based off `debian:10-slim`. This means
`yum` is no longer to be used for installing packages -- `apt-get`
should be used instead.