fix(java): implemented interface is not registered
When extending a "jsii well-known type", additional jsii interfaces
implemented were not registered with the kernel, because of a flaw in
how interface discovery works. The `@Jsii` annotation is `@Inherited`,
meaning that a sub-class of an annotated type is annotated (this is
desirable)... However, this means the fact the annotation is present on
a type does not imply it _is_ a "jsii well-known type"... Instead, we
must check whether the annotation is _declared_ by the current type (if
not, then the annotation is inherited from a parent class).

Fixes #2951