fix(python): unable to implement additional interfaces
When a type extends a jsii class, any additional interfaces implemented
via the `@jsii.implements` annotation were not properly registered, and
overrides were not properly discovered.

This was the combination of a code-generation bug, which caused the type
passed to the `create` kernel method to be that of the jsii class,
instead of the dynamic class (`self.__class__`); and a kernel bug, which
caused those inheritance graphs to completely skip overrides detection.

Fixes #2963