chore: upgrade baseline node requirement to >=12.7.0 (#2946)
Node 10.x has been end-of-lifed on `2021-04-30`, and it is more than
time to drop support for it entirely. Per our support policy, only
current releases of runtimes are supported, and hence we are at liberty
to make this change without incurring a major version bump.

BREAKING CHANGE: the minimum supported version of `node` changed from
`10.3.0` to `12.7.0`. We recommend using `16.x` if possible.