fix(java): log messages may not flush by app exit
In some cases, messages sent by JavaScript using `console.log` and
similar may not be completely flushed by the time the Java thread that
forwards that data is closed. This is because the thread might not have
been given a chance to run when `this.stop` is set to true, which makes
it exit immediately.

This change adds a new step that triggers once `this.stop` becomes
`true` that will flush the `inputStream` until EOF in a blocking way.
When this happens, the child process has already been requested to exit,
and so EOF should be reached relatively quickly.

Fixes #3009