chore: remove `jest-expect-message`
The combination of:

* jest 27
* ts-jest
* jest-expect-message

Leads to any test that wants to report an error message failing with
the error:

```
TypeError: matcherResult.message is not a function
```

(ref: https://github.com/mattphillips/jest-expect-message/issues/39)

The result is that the actual error is swallowed, and this is very
unhelpful.

Since we only used to use `jest-expect-message` in one place anyway,
remove it.