fix: modules not built sequentially (#3045)
For Go, JS, and Python, the `OneByOneBuilder` is used to build the modules
sequentially. The modules are first sorted topologically, then passed into the
builder to build one at a time. However, the package builds are kicked off
asynchronously, leading to parallel builds and potential build issues when a
local package depends on another.