feat(jsii): added warnings for usage of deprecated elements  (#3051)
General idea:

First, `DeprecationWarningsInjector.process()` consumes the assembly. For each complex type in the assembly, it generates a function responsible for calling the `print` function if the type is deprecated. For each struct property or enum member, it also generates a call to the function corresponding to the type of that property or member. This creates a call tree that mirrors the type tree (including recursive functions that mirror recursive types).

Then `Transformer` traverses the AST and for each parameter of a method, it injects into that method a call to the function corresponding to the type of that parameter.

For example, for the types:

```ts

/** @deprecated Use something else */
interface Foo {
  readonly bar: Bar;
}

interface Bar {
  readonly bar: Bar;
}

```

we get:

```ts
module_name_Foo(p) {
  if (p) { // <-- this if is to make sure we only print a warning if the user assigned it a value
    print(p, "module_name.Foo", "Use something else");
    module_name_Bar(p.bar);
  }
}

module_name_Bar(p) {
  if (p) {
    module_name_Bar(p.bar);
  }
}
```

Then, if we have a method like:

```ts
/** @deprecated Do something else */
doSomething(foo: Foo, bar: Bar) {}
```

the following calls will be injected into `doSomething`:

```ts
print("", "module_name.SomeClass#doSomething", "Do something else");
module_name_Foo(foo);
module_name_Bar(bar);
```
---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0