chore(superchain): add NOTICE file at root of image
Attempt to clarify attribution for the software included in the image by
generating a NOTICE file that contains pointers to the LICENSE/NOTICE or
similar document for everything that is included in the image.

This leverages the `/usr/share/doc/*/copyright` file wherever possible
(this is the majority of packages), and hand-maintained references for
the handful of packages that either do not get installed by `apt-get` or
that do not come with a `copyright` file.