fix(rosetta): breaks when given a lot of snippets
Rosetta started breaking recently (the process just exiting witout any error),
presumably because we increased the number of compiling code snippets and
hitting some undocumented limit in the worker message API.

Upon closer inspection the `ts.Diagnostic` objects we were sending over the
wire from the worker to the main process were holding on to a `ts.SourceFile`
object, and by extension to the entire parse tree and internal TypeScript
objects.

Instead, introduce a new type of object (`RosettaDiagnostic`) which is
a simple data holder with a couple of fields, and render the original
TypeScript diagnostics down to these and send these over the wire.

This should improve memory usage, and appears to make the rosetta
process finish again.