feat(rosetta): switch to continuously fed worker pool
Previously, Rosetta would divide all the examples to translate into `N` equally
sized arrays, and spawn `N` workers to translate them all.

Experimentation shows that the time required to translate samples is very
unequally divided, and many workers used to be idle for half of the time,
hurting throughput.

Switch to a model where we have `N` workers, and we constantly feed them a
small amount of work until all the work is done. This keeps all workers
busy until the work is complete, improving the throughput a lot.

On my machine, improves a run of Rosetta on the CDK repository
with 8 workers from ~30m to ~15m.