fix(pacmak): '*/' not escaped in JavaDocs
If any of the generated JavaDoc contains '*/', that character combination will
terminate the doc comment, throwing the Java compiler for a loop.

The character combination needs to be escaped. This change makes examples found
in the jsii assembly render the same as sample code found in running (MarkDown)
text, in doc blocks and READMEs.