fix(jsii): deprecation warnings erroneously warn for duplicate enum values (#3105)
The deprecation warnings feature will warn if an enum value is used which is
equal to one to a deprecated enum member. Unfortunately, in the case of an enum
with duplicate values, there is no way to tell whether the deprecated member was
used (or not). The current behavior emits warnings even in cases where the
active/undeprecated member is being used.

This change flips the behavior to mute the notification in the case where the
value used is a known duplicate value. This results in erring on the side of
false negatives (not catching deprecated usage) over false positives (warning in
the case nothing's wrong).

fixes #3103



---

By submitting this pull request, I confirm that my contribution is made under the terms of the [Apache 2.0 license].

[Apache 2.0 license]: https://www.apache.org/licenses/LICENSE-2.0