fix(jsii): unexpected deprecated warnings when a deprecated interface is extended
It is common in the AWS CDK for non-deprecated interfaces to extend
deprecated interfaces.
Currently, when this occurs and the non-deprecated subclass is
referenced, a warning is generated notifying the deprecation of its
superclass.

Modify the implementation such that the warning is produced only when
a property of the deprecated interface is referenced.

In the case when a deprecated interface with no required properties is
used without any of its properties set, a warning will not be generated.
This should occur rarely and the trade off is acceptable.

fixes #3111