fix(rosetta): crashes on outdated tablet files in a package
A package may accidentally (or on purpose) contain a `.jsii.tabl.json`
file, containing pretranslated examples.

When an assembly is added for snippet discovery, this tablet will
be loaded implicitly to save live conversion time.

However, if the tablet file is old, or not even a tablet file, this
will stop rosetta with an error. Catch that error and continue
anyway.

Fixes an issue in Construct Hub.