fix(pacmak): re-introduce parallelism into OneByOneBuilder
In #3045, the `OneByOneBuilder` was switched to build all packages
in series (instead of all in parallel), because there would
be race conditions in case of inter-package dependencies.

This is correct, but leaves a bunch of possible parallelism on the
table that is notably blowing up pack times for Python.

Re-introduce a (limited) form of parallelism by retaining the
sets of mutually independent packages, as toposorted, and
doing those in parallel.