fix(pacmak): add 'silly' loglevel for command output
Pacmak outputs both the command-line commands as well as stdout/stderr
for those commands under the `verbose` loglevel.

That's too much at one log level: I definitely want to know about the
commands without getting spammed with all the command outputs.

Introduce a new loglevel, `silly`, and move the command output to
that level.

Also remove a log line from `npmPack` that is redundant with the
log line that prints the command that's going to be executed.