fix(rosetta): `infuse` creates untranslated examples
The new `infuse` feature replicates examples from READMEs to types.

However, because of caching, example locations have become part of their
identifying key. Because an example has moved to a different location,
it now has a different key, and because that key is not in the tablet
yet the example will count as untranslated (requiring live conversion
during `pacmak` time, and increasing the pack time).

When we copy an example to a different type, also create a copy at
the new type location in the tablet, so that it will count as
pretranslated.